centos latest is now 8 and building with no cached layers fails
centos latest is now 8 and building with no cached layers fails.

we have rpms that are specific to centos7. We need to exlicitly call out centos 7 as the base image now that 8 has been released.

In the future we also need to sort out the python versions and equivilant epel repo location for centos 8.